refactor(cli): migrate remaining root-level tests to colocated structure
Migrate tests from src/__tests__/ to command-level colocated tests:
- compose-scope.test.ts → run/__tests__/index.test.ts
- compose-validation.test.ts → compose/__tests__/index.test.ts
- compose-versioning.test.ts → compose/__tests__/index.test.ts
- runner-validation.test.ts → compose/__tests__/index.test.ts
- index.test.ts (placeholder) → deleted

All tests already followed CLI testing patterns (command.parseAsync(),
MSW for API mocking, vi.stubEnv() for config).
