--- turbo/apps/cli/src/__tests__/index.test.ts
@@ -1,7 +1,0 @@
-import { test, expect, describe } from "vitest";
-
-describe("CLI Tests", () => {
-  test("should run in test environment", () => {
-    expect(typeof process.version).toBe("string");
-  });
-});
